+++ CHANGELOG.md
@@ -8,6 +8,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: `TextInput`: Add `rounded` boolean prop, defaults to `false`.
 - Feature: `TextInputWithButton`: Add `inset` prop to place the button inside the input.
 - Feature: `Overlay`: Add `position` prop, defaults to `absolute`.
+- Feature: Add `HorizontalCrowdfundingCard` molecule component.
+- Feature: Add `FlexWrapper` layout component.
 
 ## [4.0.2] - 2021-07-27